+++ workers/ohsome_quality_analyst/api/api.py
@@ -36,6 +36,7 @@ from ohsome_quality_analyst.definitions import (
     get_attribution,
     get_dataset_names,
     get_fid_fields,
+    get_indicator_descriptions,
     get_indicator_names,
     get_layer_keys,
     get_report_names,
@@ -325,6 +326,14 @@ async def list_fid_fields():
     """List available fid fields for each dataset."""
     response = empty_api_response()
     response["result"] = get_fid_fields()
+    return response
+
+
+@app.get("/indicator-description")
+async def get_indicator_description(IndicatorName: str):
+    """Get description of a specific indicator."""
+    response = empty_api_response()
+    response["result"] = get_indicator_descriptions(IndicatorName)
     return response