+++ workers/tests/unittests/test_definitions.py
@@ -89,3 +89,7 @@ class TestDefinitions(unittest.TestCase):
             layers,
             ("minimal",),
         )
+
+    def test_get_indicator_descriptions(self):
+        description = definitions.get_indicator_descriptions("Minimal")
+        self.assertEqual(description, "An minimal Indicator for testing purposes.\n")